requirements.txt
# Core dependencies for eSim Platform Integration
numpy>=1.21.0
matplotlib>=3.4.0
scipy>=1.7.0

# Circuit analysis
PySpice>=1.5

# Configuration management
pyyaml>=5.4.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Utilities
python-dotenv>=0.19.0
